requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.3b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==5.0.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
httpie==1.0.3
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.1.0
packaging==20.0
pathspec==0.7.0
pkg-resources==0.0.0
pluggy==0.13.1
pprint==0.1
prometheus-client==0.7.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.2
pytest-black==0.3.7
pytest-cov==2.8.1
python-dateutil==2.8.1
regex==2020.1.8
requests==2.22.0
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
uWSGI==2.0.18
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
xmltodict==0.12.0
yapf==0.29.0
zipp==0.6.0
